pyproject.toml
[project]
name = "langunan-gpt"
version = "0.1.0"
description = "LangunanGPT - Chainlit POC"
requires-python = ">=3.10"

dependencies = [
  "chainlit>=1.0.0",
  "openai>=1.0.0",
  "python-dotenv"
]
